requirements.txt
langchain>=0.1.0
langchain-openai
langchain_core
langchain-community
langchainhub
openai
ragas
faiss-cpu>=1.7.0
PyMuPDF
pandas
unstructured
datasets
chainlit==0.7.700
cohere==4.37
tiktoken==0.5.2
python-dotenv==1.0.0
requests
beautifulsoup4
reportlab
datasets